Fix potential issue with ignoring problems
A region highlight was not cleared when the problem was ignored.
I could not reproduce this issue but suspect it was because of a race
between a dummy insertion operation in ignoreProblem() (which was
triggering on_modified() in LanguageToolListener) and the subsequent
command which was clearing the region. I swapped the order of those
instructions as a precaution.